DbConnection no longer abstract
